Remove default footer and propose a react component to replace it.
