+++ docusaurus.config.ts
@@ -64,7 +64,7 @@ const config: Config = {
       title: "",
       logo: {
         alt: "QuantStack Logo",
-        src: "img/quantstack/logo-website.svg"
+        src: "img/quantstack/logo-website.svg",
       },
 
       items: [
@@ -131,96 +131,10 @@ const config: Config = {
       ],
     },
 
-    /*footer: {
-      links: [
-        {
-          title: " ",
-          items: [
-            {
-              label: " ",
-              className: "astronaut-footer",
-              href: "/home",
-            },
-          ],
-        },
-        
-        {
-          title: "Follow QuantStack on",
-          items: [
-            {
-              label: "GitHub",
-              className: "github-icon",
-              href: "https://github.com/QuantStack",
-            },
-            {
-              label: "LinkedIn",
-              className: "linkedin-icon",
-              href: "https://www.linkedin.com/company/quantstack/mycompany/",
-            },
-            {
-              label: "X",
-              className: "x-icon",
-              href: "https://twitter.com/QuantStack",
-            },
-          ],
-        },
-        {
-          title: "Menu ",
-          items: [
-            {
-              label: "Home",
-              href: "/home/",
-            },
-            {
-              label: "Projects",
-              href: "/projects/",
-            },
-            {
-              label: "Services",
-              href: "/services/",
-            },
-            {
-              label: "About",
-              href: "/about/",
-            },
-            {
-              label: "Careers",
-              href: "/careers/",
-            },
-            {
-              label: "Blog",
-              href: "/blog/",
-            },
-          ],
-        },
-        
-
-        {
-          title: "QuantStack office ",
-          items: [
-            {
-              html: `
-              
-                  <div class="div .address">
-                      16, avenue Curti <br/>
-                      94100 Saint-Maur-des-Fossés <br/>
-                      France
-                  </div>
-               
-                    `,
-            },
-          ],
-        },
-      ],
-
-      copyright: copyright,
-    },*/
-    
-      colorMode: {
-        defaultMode: 'light',
-        disableSwitch: true,
-        respectPrefersColorScheme: false,
-      
+    colorMode: {
+      defaultMode: "light",
+      disableSwitch: true,
+      respectPrefersColorScheme: false,
     },
 
     prism: {
