+++ src/pages/about.tsx
@@ -2,12 +2,14 @@ import Layout from "@theme/Layout";
 import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
 import {About} from "../components/about";
 import BrowserOnly from "@docusaurus/BrowserOnly";
+import Footer from "../components/footer/Footer";
 
 export default function AboutPage(): JSX.Element {
   const { siteConfig } = useDocusaurusContext();
   return (
     <Layout>
       <BrowserOnly>{() => <About />}</BrowserOnly>
+      <Footer/>
     </Layout>
   );
 }
