+++ src/pages/blog.tsx
@@ -2,6 +2,7 @@ import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
 import Layout from "@theme/Layout";
 import BlogsComponent from "../components/blog";
 import blogpostDetails from "@site/src/components/blog/blogpostsDetails.json";
+import Footer from "../components/footer/Footer";
 
 
 export default function BlogPage(): JSX.Element {
@@ -9,6 +10,7 @@ export default function BlogPage(): JSX.Element {
   return (
     <Layout>
       <BlogsComponent blogpostsDetails={blogpostDetails} />
+      <Footer/>
     </Layout>
   );
 }
