+++ src/pages/careers.tsx
@@ -1,6 +1,7 @@
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
 import Layout from '@theme/Layout';
 import Careers from '../components/careers';
+import Footer from "../components/footer/Footer";
 
 
 export default function CareersPage(): JSX.Element {
@@ -8,6 +9,7 @@ export default function CareersPage(): JSX.Element {
   return (
     <Layout>
       <Careers />
+      <Footer/>
     </Layout>
   );
 }
