+++ src/pages/home.tsx
@@ -2,6 +2,7 @@ import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
 import Layout from "@theme/Layout";
 import Home from "../components/home/Home";
 import BrowserOnly from "@docusaurus/BrowserOnly";
+import Footer from "../components/footer/Footer";
 
 export default function HomePage(): JSX.Element {
   const { siteConfig } = useDocusaurusContext();
@@ -9,6 +10,7 @@ export default function HomePage(): JSX.Element {
     <Layout>
       <main>
       <BrowserOnly>{() => <Home />}</BrowserOnly>
+      <Footer/>
       </main>
     </Layout>
   );
