+++ src/pages/index.tsx
@@ -2,7 +2,7 @@ import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
 import Layout from "@theme/Layout";
 import Home from "../components/home/Home";
 import BrowserOnly from "@docusaurus/BrowserOnly";
-import Footer from "../components/home/Footer";
+import Footer from "../components/footer/Footer";
 
 export default function QSWebsite(): JSX.Element {
   const { siteConfig } = useDocusaurusContext();
