+++ src/pages/services.tsx
@@ -2,6 +2,7 @@ import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
 import Layout from "@theme/Layout";
 import Services from "../components/services";
 import BrowserOnly from "@docusaurus/BrowserOnly";
+import Footer from "../components/footer/Footer";
 
 export default function ServicesPage(): JSX.Element {
   const { siteConfig } = useDocusaurusContext();
@@ -10,6 +11,7 @@ export default function ServicesPage(): JSX.Element {
       <BrowserOnly>
      {() => <Services/>}
       </BrowserOnly>
+      <Footer/>
     </Layout>
   );
 }
